Role init

+++ ansible_dev/config_handler/config.py
@@ -27,8 +27,9 @@ class ConfigHandler(object):
                 os.makedirs(path)
             self.find_files_in_current_package()
             for files in self._config_files:
-                shutil.copy(files, path)
-            # Update Config files of sysytem to writable files
+                if not os.path.exists(os.path.join(path, os.path.basename(files))):
+                    shutil.copy(files, path)
+            # Update Config files on this.object to writable files location
             filenames = []
             for files in self._config_files:
                 filenames.append(os.path.basename(files))
@@ -37,7 +38,7 @@ class ConfigHandler(object):
             for files in filenames:
                 self._config_files.append(os.path.join(self._path, files))
         except (OSError, IOError) as e:
-            print ('Excepion received : %s' % e)
+            print ('creating config file at user editable.Excepion received : %s' % e)
 
     def load_config_from_path(self, file_path):
         config_file_abspath = os.path.abspath(file_path)
@@ -60,16 +61,16 @@ class ConfigHandler(object):
 
         for section in kwargs:
             if section in old_sections:
-                old_sections.update(kwargs[section])
+                for key in kwargs[section]:
+                   old_config[section][key] = kwargs[section][key]
             else:
                 old_config[section] = kwargs[section]
         
         try:
              with open(filename_path, 'w') as f:
                  old_config.write(f)
-                 if new_file_loc is not None:
-                     old_config.write(new_file_loc)
-                 
+             if new_file_loc is not None:
+                 shutil.copy(filename_path, new_file_loc)
         except (OSError, IOError) as e:
             print ("Unable to modify cfg file : %s" % filename_path)
 
@@ -84,6 +85,17 @@ class ConfigHandler(object):
             filename = os.path.basename(file_path)
             if fnmatch.fnmatch(filename, 'ansible.cfg'):
                 self.update_config(file_path, new_file_path, **kwargs)
+
+    def get_value(self, filename, arg_section, key):
+        for file_path in self._config_files:
+            filen = os.path.basename(file_path)
+            if fnmatch.fnmatch(filen, filename):
+                c_parser = self.load_config_from_path(file_path)
+                sections = c_parser.sections()
+                for section in sections:
+                    if arg_section == section:
+                        return c_parser[section].get(key)
+        return None
 
 if __name__ == "__main__":
     an_config = ConfigHandler()
